+++ knwl/cli/cli.py
@@ -8,6 +8,7 @@ import json
 from knwl.cli.config_app import config_app
 from knwl.cli.info_app import info_app
 from knwl.cli.log_app import log_app
+from knwl.cli.collect_app import collect_app
 from knwl.cli.graph_app import graph_app, similar_nodes, find_nodes
 
 from rich.console import Console
@@ -23,6 +24,7 @@ app.add_typer(config_app, name="config", context_settings={"obj": K})
 app.add_typer(info_app, name="info", context_settings={"obj": K})
 app.add_typer(log_app, name="log", context_settings={"obj": K})
 app.add_typer(graph_app, name="graph", context_settings={"obj": K})
+app.add_typer(collect_app, name="collect", context_settings={"obj": K})
 
 
 @app.command(